fixed search apartments

+++ AirBNB/Models/Apartment.cs
@@ -162,7 +162,7 @@ namespace AirBNB.Models
             return ds.getApartmentByID(id);
         }
 
-        public List<Apartment> getAllApartmentsBySearch(string keyword, string from, string to, int minP, int maxP, double minD, double maxD, int beds, int rating)
+        public List<Apartment> getAllApartmentsBySearch(string keyword, string from, string to, int minP, int maxP, double minD, double maxD, int beds, double rating)
         {
             DataServices ds = new DataServices();
             return ds.getAllApartmentsBySearch(keyword, from, to, minP, maxP, minD, maxD, beds, rating);
